Updated code from the merge conflicts.
